Move BSF graph to Insights page and add interop graph placeholder
Relocate the Browser-Specific Failures graph from the homepage to the Insights page.
The homepage slot is replaced with a placeholder for the upcoming feature-level
interoperability graph.
